requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
Flask>=2.3.0
Werkzeug>=2.3.0
Flask-Login>=0.6.3
Flask-SQLAlchemy>=3.1.1
SQLAlchemy>=2.0.0
onvif-zeep>=0.2.12
tqdm>=4.66.0
ffmpeg-python>=0.2.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.8.0
seaborn>=0.13.0
pytest>=7.4.0
pyyaml>=6.0.0
